Update DocumentScannerGUI.py
FIxed bug with the citation map where closing the program wouldnt finish the program and it would loop in the background. 

Added new function close_graph_and_back for the back button in sho_citation_graph_gui 

Changed extract citations that way it would search for a works cited page and find specific text file names in that page.